requirements.txt
english-words
Pillow
tkmacosx

